+++ src/main/java/ru/mystamps/web/service/UserService.java
@@ -28,5 +28,5 @@ public interface UserService {
 	int countRegistrationRequestByActivationKey(String activationKey);
 	void registerUser(ActivateAccountDto dto);
 	User findByLogin(String login);
-	int countByLogin(String login);
+	long countByLogin(String login);
 }
